Move utility functions to another file
